polymarket_hft/requirements.txt
websockets>=12.0,<14.0
aiohttp>=3.9,<4.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
py-clob-client>=0.5.0
